Added Shopping Cart Page with Summary Component. Hooked up Redux to Cart.
Created Shopping Cart page when users click on the "add to cart" button.
This will allow them to keep track of their items in the cart when visiting
the page. Created the cart item component and summary cart card as well.
These two components were hooked into Redux to show the correct subtotal
and total when removing and adding items.